requirements.txt
numpy==1.20.2
PyOpenGL==3.1.5
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
